+++ apps/antalmanac/src/actions/AppStoreActions.ts
@@ -15,8 +15,8 @@ import type {
     RepeatingCustomEvent,
     ScheduleCourse,
     ShortCourseSchedule,
-    WebsocSection,
 } from '@packages/antalmanac-types';
+import type { WebsocSection } from '@packages/anteater-api/types';
 import { TRPCClientError } from '@trpc/client';
 import type { PostHog } from 'posthog-js/react';
