+++ apps/antalmanac/src/components/Calendar/CourseCalendarEvent.tsx
@@ -12,7 +12,8 @@ import { formatTimes } from '$stores/calendarizeHelpers';
 import { useTimeFormatStore } from '$stores/SettingsStore';
 import { Delete, Search } from '@mui/icons-material';
 import { Chip, IconButton, Paper, Tooltip, Button, Box } from '@mui/material';
-import { CustomEventId, WebsocSectionFinalExam } from '@packages/antalmanac-types';
+import { CustomEventId } from '@packages/antalmanac-types';
+import { WebsocSectionFinalExam } from '@packages/anteater-api/types';
 import { usePostHog } from 'posthog-js/react';
 import { useEffect, useRef } from 'react';
 import { Event } from 'react-big-calendar';
