+++ apps/antalmanac/src/components/RightPane/SectionTable/PrereqTree.tsx
@@ -1,6 +1,6 @@
 import { CourseInfo } from '$components/RightPane/SectionTable/CourseInfo/CourseInfoBar';
 import { Button, Popover, useTheme } from '@mui/material';
-import { Prerequisite, PrerequisiteTree } from '@packages/antalmanac-types';
+import { Prerequisite, PrerequisiteTree } from '@packages/anteater-api/types';
 import { FC, useState } from 'react';
 
 import './PrereqTree.css';
