+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/DetailsCell.tsx
@@ -1,7 +1,6 @@
-import { Box, SxProps } from '@mui/material';
-import { WebsocSectionType } from '@packages/antalmanac-types';
-
 import { TableBodyCellContainer } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/TableBodyCellContainer';
+import { Box, SxProps } from '@mui/material';
+import { WebsocSectionType } from '@packages/anteater-api/types';
 
 const SECTION_COLORS = {
     Act: { color: '#c87137' },
