+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/SyllabusCell.tsx
@@ -1,7 +1,6 @@
-import { WebsocSectionStatus } from '@packages/antalmanac-types';
-import { Link } from 'react-router-dom';
-
 import { TableBodyCellContainer } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/TableBodyCellContainer';
+import { WebsocSectionStatus } from '@packages/anteater-api/types';
+import { Link } from 'react-router-dom';
 
 interface SyllabusCellProps {
     webURL: WebsocSectionStatus;
