+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTablePopover/PastSyllabiPopover.tsx
@@ -12,7 +12,7 @@ import {
     Skeleton,
     Typography,
 } from '@mui/material';
-import type { WebsocSyllabiResponse } from '@packages/antalmanac-types';
+import type { WebsocSyllabiResponse } from '@packages/anteater-api/types';
 import Link from 'next/link';
 import { useEffect, useMemo, useState } from 'react';
