+++ apps/antalmanac/src/lib/download.ts
@@ -3,7 +3,7 @@ import buildingCatalogue from '$lib/locations/buildingCatalogue';
 import { getDefaultTerm, termData } from '$lib/termData';
 import AppStore from '$stores/AppStore';
 import { openSnackbar } from '$stores/SnackbarStore';
-import type { HourMinute } from '@packages/antalmanac-types';
+import type { HourMinute } from '@packages/anteater-api/types';
 import { saveAs } from 'file-saver';
 import { createEvents, type EventAttributes } from 'ics';
