+++ apps/antalmanac/src/lib/grades.ts
@@ -1,6 +1,5 @@
-import { GE } from '@packages/antalmanac-types';
-
 import trpc from '$lib/api/trpc';
+import { GE } from '@packages/anteater-api/types';
 
 export interface GradesProps {
     averageGPA: number | null;
