+++ packages/types/src/index.ts
@@ -8,4 +8,3 @@ export * from './search';
 export * from './websoc';
 export * from './planner';
 export * from './notification';
-export * from '@packages/anteater-api/types';
